Move code from `NightscoutMongoCommand` into the main `NightscoutCommand` class
