+++ src/main/java/com/dongtronic/diabot/Main.kt
@@ -10,7 +10,7 @@ import com.dongtronic.diabot.platforms.discord.commands.diabetes.EstimationComma
 import com.dongtronic.diabot.platforms.discord.commands.info.InfoCommand
 import com.dongtronic.diabot.platforms.discord.commands.misc.*
 import com.dongtronic.diabot.platforms.discord.commands.nightscout.NightscoutAdminCommand
-import com.dongtronic.diabot.platforms.discord.commands.nightscout.NightscoutMongoCommand
+import com.dongtronic.diabot.platforms.discord.commands.nightscout.NightscoutCommand
 import com.dongtronic.diabot.platforms.discord.commands.quote.QuoteCommand
 import com.dongtronic.diabot.platforms.discord.commands.rewards.RewardsCommand
 import com.dongtronic.diabot.platforms.discord.listeners.*
@@ -79,7 +79,7 @@ object Main {
 
                 // BG
                 ConvertCommand(bgCategory),
-                NightscoutMongoCommand(bgCategory),
+                NightscoutCommand(bgCategory),
 
                 // Utility
                 PingCommand(utilitiesCategory),
